osd: fix decoding of legacy (v2) coll_t
It was u8, not int.

+++ src/osd/osd_types.h
@@ -290,7 +290,7 @@ public:
       }
 
       case 2: {
-	int type;
+	__u8 type;
 	pg_t pgid;
 	snapid_t snap;
